Make bootstrap optional on a package level
According to the readme "Optionally uses bootstrap.css (v 3.x.x) for styling of some elements (although the component is fully functional without it and there is a flag to turn off the dependency)."

I was however not able to find that flag. 

With the above change it should install without even pulling in the bootstrap package, just by specifying --no-optional on npm install.